Move `HeadingLevelIcon` component to a separate file
